Add NUGET_API_KEY to publish.yml for NuGet publishing
The changes in the `publish.yml` file involve modifying the job steps for packing and publishing a NuGet package. Specifically, an environment variable `NUGET_API_KEY` has been added to the `Publish to NuGet` step. This variable is set using the secret `NUGET_API_KEY`, which is likely used for authentication when pushing the NuGet package to the NuGet repository.